[components] First commit, all polished and set to start coding Glacier UI!
Hello people :)
I took inspiration from the current qtquickcontrols-tizen repo and
built the repo for the new Nemo components for Glacier UI!
So, first of all, a big thank you to Tomasz Olszak (nickname: tolszak)
and Jarosław_Staniek (nickname: staniek) for their awesome work on
Qt for Tizen!

I changed quite a lot of stuff from qtquickcontrols-tizen repo
in the attempt to improve the overall quality and make things
easier for the people who will contribute to building Glacier
components. I hope I succeeded at that task :)

For those who don't know what Glacier is:
It is the new UI for Nemomobile, you can find some mockup here:
http://play.qwazix.com/grog/?p=344
Yes, it will be as awesome as it looks from the mockups! :D

If you want to be part of the development please do not hesitate
to come talk to me (faenil) at #nemomobile IRC channel (FreeNode server)
and...let's make it happen!

Cheers,
faenil
(Andrea Bernabei)