[Bugfix] Bugfixed the generator
The class generator used to provide wrong values for
properties that have a default value. This bugfix
also includes the capability to restore a default
value when switching from a theme providing an
overridden value to a theme providing no value.
Before, since there were no value, the old value
from the old theme was kept.